Combine the KOKKOS bug fix branch with the KOKKOS test coverage branch on
top of upstream develop.  Where both branches fixed the same defect the
better solution was kept:

- temp/region bias detection in compute temp/sphere, temp/asphere and
  temp/body: prefix match on the style name (also covers temp/region/eff)
- compute inertia/kk: test the per-atom flags (radius, ellipsoid, line,
  tri, body) instead of the atom style names
- ML-SNAP grid computes (sna/grid, sna/grid/local, gaussian/grid/local):
  the full rework (type-indexed element views, group mask, host fallback,
  Kokkos-owned grid storage, neighbor overflow retry) supersedes the
  smaller indexing fix
- DSF and Wolf pair styles: the special-bonds fix is identical in both
  branches; the self energy is tallied in a device kernel instead of a
  host loop, which also removes the need for the host charge sync; the
  single-precision safe EPSILON floor of born/coul/dsf/cs is applied to
  born/coul/wolf/cs as well
- fix neigh/history/kk: identical fixes; the restart unpacking variant that
  also updates maxexchange is kept
- pair brownian/kk: host random number generator support (KOKKOS_DEBUG_RNG)
  plus the rejection of full neighbor lists
- pair multi/lucy/rx/kk: table and cutoff refresh plus the per-atom energy
  tally of the density term
- atom style ellipsoid/kk: the superellipsoid option is rejected once, in
  process_args()
- min cg/kk and sd/kk: one force sync after the line search
- kokkos.cpp citation: "--" page range, as BibTeX expects

Additionally, set Pair::hybrid_index in every place hybrid sub-styles are
created (PairHybrid::read_restart, hybrid/scaled, hybrid/scaled/kk), so the
restart-stable contact history fix id also holds when a hybrid pair style
is read back from a restart file.

+++ src/pair.h
@@ -250,7 +250,7 @@ class Pair : protected Pointers {
  protected:
   int instance_me;      // which Pair class instantiation I am
   int instance_index();    // position within the current simulation, for internal fix ids
-  int special_lj[4];    // copied from force->special_lj for Kokkos
+  double special_lj[4];    // copied from force->special_lj for Kokkos
 
   // pair_modify settings
   int offset_flag, mix_flag;    // flags for offset and mixing
